feat: 자기소개서, 스펙 정보 컴포넌트 제작
